+++ src/components/common/InputField/index.tsx
@@ -1,5 +1,31 @@
+/* eslint-disable react/jsx-props-no-spreading */
+/* eslint-disable react/require-default-props */
 import React from 'react';
 import styled from '@emotion/styled';
+
+interface InputFieldProps extends React.InputHTMLAttributes<HTMLInputElement> {
+  bgColor?: string;
+  radius?: string;
+  width?: string;
+}
+
+const InputField: React.FC<InputFieldProps> = ({
+  placeholder,
+  value,
+  bgColor = '#ffffff',
+  radius = '8px',
+  width = '100%',
+  ...rest
+}) => (
+  <StyledInput
+    placeholder={placeholder}
+    value={value}
+    bgColor={bgColor}
+    radius={radius}
+    width={width}
+    {...rest}
+  />
+);
 
 const StyledInput = styled.input<{
   bgColor: string;
@@ -23,37 +49,5 @@ const StyledInput = styled.input<{
     border-color: #a0a0a0;
   }
 `;
-
-interface InputFieldProps {
-  placeholder?: string;
-  value?: string;
-  bgColor?: string;
-  radius?: string;
-  width?: string;
-}
-
-const InputField: React.FC<InputFieldProps> = ({
-  placeholder,
-  value,
-  bgColor = '#ffffff',
-  radius = '8px',
-  width = '100%',
-}) => (
-  <StyledInput
-    placeholder={placeholder}
-    value={value}
-    bgColor={bgColor}
-    radius={radius}
-    width={width}
-  />
-);
-
-InputField.defaultProps = {
-  placeholder: '입력해 주세요',
-  value: '',
-  bgColor: '#ffffff',
-  radius: '8px',
-  width: '100%',
-};
 
 export default InputField;
